Home work week 4/requirements.txt
streamlit
opencv-python
numpy
Pillow
hugchat
